Break into two projects, a cli tool (framework) and a task (core and framework)
