+++ src/SignCheck/SignCheckTask/src/LICENSE.TXT
