+++ src/SignCheck/SignCheckTask/src/SignCheckResources.resx
